Split up CLI commands into separate modules.
