+++ src/case_rate/cli/__init__.py
@@ -1,0 +1,1 @@
+from .cli import main
